refactor(io): regroup serializers into data/model/raster domains
safetensors plumbing (tensors, bytes) moves under model/ with its only
users; json/path stay at the io root as generic plumbing shared across
domains.
