+++ cli/src/commands/plot/run.rs
@@ -4,9 +4,9 @@ use crate::path::PathExt;
 use clap::Args;
 use nrn::data::Dataset;
 use nrn::data::scalers::ScalerMethod;
-use nrn::io::checkpoint::CheckpointArchive;
-use nrn::io::gif::GifWriter;
-use nrn::io::run::TrainingRun;
+use nrn::io::model::checkpoint::CheckpointArchive;
+use nrn::io::model::run::TrainingRun;
+use nrn::io::raster::gif::GifWriter;
 use nrn::model::Predictor;
 use nrn::plot::{ConsoleConfig, Figure};
 use nrn::task::Task;
