+++ cli/src/commands/train/start.rs
@@ -6,9 +6,9 @@ use crate::path::PathExt;
 use clap::Args;
 use nrn::activations::RELU;
 use nrn::data::Dataset;
-use nrn::io::hyperparams::HyperParametersRecord;
-use nrn::io::network::NetworkConfig;
-use nrn::io::run::{TrainingMeta, TrainingRun};
+use nrn::io::model::hyperparams::HyperParametersRecord;
+use nrn::io::model::network::NetworkConfig;
+use nrn::io::model::run::{TrainingMeta, TrainingRun};
 use nrn::model::{LayerPlan, NeuralNetwork, NeuronLayerSpec, Predictor};
 use nrn::task::Task;
 use nrn::training::Callbacks;
