+++ core/src/io/data/mod.rs
@@ -1,0 +1,3 @@
+pub mod classes;
+pub mod dataset;
+pub mod instance;
