+++ core/src/io/mod.rs
@@ -1,22 +1,9 @@
 pub mod bytes;
-pub mod checkpoint;
-pub mod classes;
-pub mod data;
-pub mod hyperparams;
-pub mod instance;
 pub mod json;
-pub mod network;
-pub mod optimizer;
 pub mod path;
-pub mod predictor;
-pub mod run;
-pub mod scalers;
-pub mod scheduler;
-pub mod task;
-pub mod tensors;
 
-// Figure persistence — only available alongside the raster renderer.
+pub mod data;
+pub mod model;
+
 #[cfg(feature = "raster")]
-pub mod gif;
-#[cfg(feature = "raster")]
-pub mod png;
+pub mod raster;
