+++ core/src/io/raster/gif.rs
