+++ core/src/io/raster/mod.rs
@@ -1,0 +1,2 @@
+pub mod gif;
+pub mod png;
